+++ vite.config.ts
@@ -2,7 +2,7 @@ import { defineConfig } from 'vite';
 import { resolve } from 'path';
 
 export default defineConfig({
-  base: '/Minecraft-Web/',
+  base: './',
   resolve: {
     alias: {
       '@': resolve(__dirname, 'src')